requirements.txt
azure-ai-inference
sql-metadata
google-genai
mistralai>=1.0.0
openai
pandas
transformers
torch
tabulate